Refactor auth code
